Fleshed out Beam Wizard

+++ Space/Assets/Scripts/Gameplay/Weapon/Pew Pewers/WeaponScript.cs
@@ -39,7 +39,7 @@ public class WeaponScript : MonoBehaviour
 	public float attackPower;
 	public float shieldPiercing;
 	public float cooldown;
-	public float shotsBeforeCooldown;
+	public int shotsBeforeCooldown;
 	public int projectilesPerShot;
 	public float shotsPerClip;
 	public float maxReserveClips;
